Moved lm and log data to dedicated "data" folder
