+++ backend/main.py
@@ -8,13 +8,8 @@ from typing import List
 app = FastAPI()
 
 # Enable CORS
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
+# app.add_middleware(CORSMiddleware, allow_origins=["*"], allow_credentials=True, allow_methods=["*"],
+#                    allow_headers=["*"], )
 
 
 class DataPoint(BaseModel):
@@ -27,7 +22,7 @@ async def get_data():
     # Generate sample data for the last 24 hours
     data = []
 
-    for hour in range(1,25):
+    for hour in range(1, 25):
         time_str = f"{hour:02d}:00"
         value = random.randint(0, 100)  # Random value between 0-100
 
@@ -41,4 +36,5 @@ async def get_data():
 
 if __name__ == "__main__":
     import uvicorn
+
     uvicorn.run(app, host="0.0.0.0", port=8000)
